Reorganized file structure for future dev in socket server
